changes for new structure. Routes and controllers
