+++ data/boiler-type.json
@@ -1,4 +1,26 @@
-[{"id":{"$oid":"5fbbb92ffc13ae4285000000"},"desc":"papa"},
-{"id":{"$oid":"5fbbb92ffc13ae4285000001"},"desc":"papa"},
-{"id":{"$oid":"5fbbb92ffc13ae4285000002"},"desc":"non ligula pellentesque ultrices phasellus id sapien in sapien iaculis"},
-{"id":{"$oid":"5fbbb92ffc13ae4285000003"},"desc":"nulla tempus vivamus in felis eu sapien cursus vestibulum proin eu mi nulla"}]
+[
+    {
+        "id": {
+            "$oid": "5fbbb92ffc13ae4285000000"
+        },
+        "desc": "papa"
+    },
+    {
+        "id": {
+            "$oid": "5fbbb92ffc13ae4285000001"
+        },
+        "desc": "papa"
+    },
+    {
+        "id": {
+            "$oid": "5fbbb92ffc13ae4285000002"
+        },
+        "desc": "non ligula pellentesque ultrices phasellus id sapien in sapien iaculis"
+    },
+    {
+        "id": {
+            "$oid": "5fbbb92ffc13ae4285000003"
+        },
+        "desc": "nulla tempus vivamus in felis eu sapien cursus vestibulum proin eu mi nulla"
+    }
+]
