+++ data/boilers.json
@@ -1,20 +1,182 @@
-[{"name":"TrailBlazer","type":{"$oid":"5fbb1cc5fc13ae1252000000"},"_id":{"$oid":"5fbb1cc5fc13ae1252000001"}},
-    {"name":"Echo","type":{"$oid":"5fbb1cc5fc13ae1252000002"},"_id":{"$oid":"5fbb1cc5fc13ae1252000003"}},
-    {"name":"Land Cruiser","type":{"$oid":"5fbb1cc5fc13ae1252000004"},"_id":{"$oid":"5fbb1cc5fc13ae1252000005"}},
-    {"name":"Century","type":{"$oid":"5fbb1cc5fc13ae1252000006"},"_id":{"$oid":"5fbb1cc5fc13ae1252000007"}},
-    {"name":"R-Class","type":{"$oid":"5fbb1cc5fc13ae1252000008"},"_id":{"$oid":"5fbb1cc5fc13ae1252000009"}},
-    {"name":"Range Rover Sport","type":{"$oid":"5fbb1cc5fc13ae125200000a"},"_id":{"$oid":"5fbb1cc5fc13ae125200000b"}},
-    {"name":"FX","type":{"$oid":"5fbb1cc5fc13ae125200000c"},"_id":{"$oid":"5fbb1cc5fc13ae125200000d"}},
-    {"name":"Touareg","type":{"$oid":"5fbb1cc5fc13ae125200000e"},"_id":{"$oid":"5fbb1cc5fc13ae125200000f"}},
-    {"name":"Camry","type":{"$oid":"5fbb1cc5fc13ae1252000010"},"_id":{"$oid":"5fbb1cc5fc13ae1252000011"}},
-    {"name":"MDX","type":{"$oid":"5fbb1cc5fc13ae1252000012"},"_id":{"$oid":"5fbb1cc5fc13ae1252000013"}},
-    {"name":"B-Series Plus","type":{"$oid":"5fbb1cc5fc13ae1252000014"},"_id":{"$oid":"5fbb1cc5fc13ae1252000015"}},
-    {"name":"Countach","type":{"$oid":"5fbb1cc5fc13ae1252000016"},"_id":{"$oid":"5fbb1cc5fc13ae1252000017"}},
-    {"name":"Sierra 1500","type":{"$oid":"5fbb1cc5fc13ae1252000018"},"_id":{"$oid":"5fbb1cc5fc13ae1252000019"}},
-    {"name":"PT Cruiser","type":{"$oid":"5fbb1cc5fc13ae125200001a"},"_id":{"$oid":"5fbb1cc5fc13ae125200001b"}},
-    {"name":"3 Series","type":{"$oid":"5fbb1cc5fc13ae125200001c"},"_id":{"$oid":"5fbb1cc5fc13ae125200001d"}},
-    {"name":"Protege","type":{"$oid":"5fbb1cc5fc13ae125200001e"},"_id":{"$oid":"5fbb1cc5fc13ae125200001f"}},
-    {"name":"Taurus","type":{"$oid":"5fbb1cc5fc13ae1252000020"},"_id":{"$oid":"5fbb1cc5fc13ae1252000021"}},
-    {"name":"Bronco","type":{"$oid":"5fbb1cc5fc13ae1252000022"},"_id":{"$oid":"5fbb1cc5fc13ae1252000023"}},
-    {"name":"LTD Crown Victoria","type":{"$oid":"5fbb1cc5fc13ae1252000024"},"_id":{"$oid":"5fbb1cc5fc13ae1252000025"}},
-    {"name":"Corolla","type":{"$oid":"5fbb1cc5fc13ae1252000026"},"_id":{"$oid":"5fbb1cc5fc13ae1252000027"}}]
+[
+    {
+        "name": "TrailBlazer",
+        "type": {
+            "$oid": "5fbb1cc5fc13ae1252000000"
+        },
+        "_id": {
+            "$oid": "5fbb1cc5fc13ae1252000001"
+        }
+    },
+    {
+        "name": "Echo",
+        "type": {
+            "$oid": "5fbb1cc5fc13ae1252000002"
+        },
+        "_id": {
+            "$oid": "5fbb1cc5fc13ae1252000003"
+        }
+    },
+    {
+        "name": "Land Cruiser",
+        "type": {
+            "$oid": "5fbb1cc5fc13ae1252000004"
+        },
+        "_id": {
+            "$oid": "5fbb1cc5fc13ae1252000005"
+        }
+    },
+    {
+        "name": "Century",
+        "type": {
+            "$oid": "5fbb1cc5fc13ae1252000006"
+        },
+        "_id": {
+            "$oid": "5fbb1cc5fc13ae1252000007"
+        }
+    },
+    {
+        "name": "R-Class",
+        "type": {
+            "$oid": "5fbb1cc5fc13ae1252000008"
+        },
+        "_id": {
+            "$oid": "5fbb1cc5fc13ae1252000009"
+        }
+    },
+    {
+        "name": "Range Rover Sport",
+        "type": {
+            "$oid": "5fbb1cc5fc13ae125200000a"
+        },
+        "_id": {
+            "$oid": "5fbb1cc5fc13ae125200000b"
+        }
+    },
+    {
+        "name": "FX",
+        "type": {
+            "$oid": "5fbb1cc5fc13ae125200000c"
+        },
+        "_id": {
+            "$oid": "5fbb1cc5fc13ae125200000d"
+        }
+    },
+    {
+        "name": "Touareg",
+        "type": {
+            "$oid": "5fbb1cc5fc13ae125200000e"
+        },
+        "_id": {
+            "$oid": "5fbb1cc5fc13ae125200000f"
+        }
+    },
+    {
+        "name": "Camry",
+        "type": {
+            "$oid": "5fbb1cc5fc13ae1252000010"
+        },
+        "_id": {
+            "$oid": "5fbb1cc5fc13ae1252000011"
+        }
+    },
+    {
+        "name": "MDX",
+        "type": {
+            "$oid": "5fbb1cc5fc13ae1252000012"
+        },
+        "_id": {
+            "$oid": "5fbb1cc5fc13ae1252000013"
+        }
+    },
+    {
+        "name": "B-Series Plus",
+        "type": {
+            "$oid": "5fbb1cc5fc13ae1252000014"
+        },
+        "_id": {
+            "$oid": "5fbb1cc5fc13ae1252000015"
+        }
+    },
+    {
+        "name": "Countach",
+        "type": {
+            "$oid": "5fbb1cc5fc13ae1252000016"
+        },
+        "_id": {
+            "$oid": "5fbb1cc5fc13ae1252000017"
+        }
+    },
+    {
+        "name": "Sierra 1500",
+        "type": {
+            "$oid": "5fbb1cc5fc13ae1252000018"
+        },
+        "_id": {
+            "$oid": "5fbb1cc5fc13ae1252000019"
+        }
+    },
+    {
+        "name": "PT Cruiser",
+        "type": {
+            "$oid": "5fbb1cc5fc13ae125200001a"
+        },
+        "_id": {
+            "$oid": "5fbb1cc5fc13ae125200001b"
+        }
+    },
+    {
+        "name": "3 Series",
+        "type": {
+            "$oid": "5fbb1cc5fc13ae125200001c"
+        },
+        "_id": {
+            "$oid": "5fbb1cc5fc13ae125200001d"
+        }
+    },
+    {
+        "name": "Protege",
+        "type": {
+            "$oid": "5fbb1cc5fc13ae125200001e"
+        },
+        "_id": {
+            "$oid": "5fbb1cc5fc13ae125200001f"
+        }
+    },
+    {
+        "name": "Taurus",
+        "type": {
+            "$oid": "5fbb1cc5fc13ae1252000020"
+        },
+        "_id": {
+            "$oid": "5fbb1cc5fc13ae1252000021"
+        }
+    },
+    {
+        "name": "Bronco",
+        "type": {
+            "$oid": "5fbb1cc5fc13ae1252000022"
+        },
+        "_id": {
+            "$oid": "5fbb1cc5fc13ae1252000023"
+        }
+    },
+    {
+        "name": "LTD Crown Victoria",
+        "type": {
+            "$oid": "5fbb1cc5fc13ae1252000024"
+        },
+        "_id": {
+            "$oid": "5fbb1cc5fc13ae1252000025"
+        }
+    },
+    {
+        "name": "Corolla",
+        "type": {
+            "$oid": "5fbb1cc5fc13ae1252000026"
+        },
+        "_id": {
+            "$oid": "5fbb1cc5fc13ae1252000027"
+        }
+    }
+]
