+++ index.js
@@ -1,19 +1,24 @@
-
 // Declarations
 const express = require('express');
 const app = express();
+//require your controller
 const boilerController = require("./controllers/boilers");
+const technicianController = require("./controllers/technician");
+//Router
+const router = require('./routes');
+//Port
 const PORT = process.env.PORT || 4000;
 
 // API Routes
-app.use('/api/boiler', boilerController);
-app.use('/companies', require('./controllers/companies'));
+//app.use('/api/boiler', boilerController);
+//app.use('/api/technician', technicianController);
+//app.use('/companies', require('./controllers/companies'));
 
 // Require data.json
 const buildings = require('./data/building.json');
-const technicians = require('./data/technician.json');
 const boilerType = require('./data/boiler-type');
 
+app.use(router);
 // Server listener
 app.listen(PORT, () => console.log(`Server started on port ${PORT}`));
 
@@ -91,70 +96,6 @@ app.delete('/api/building/:id', (req, res) => {
         res.status(400).json({msg: `No building with id: ${req.params.id}`});
     }
 });
-
-//---------------
-// Alexis Pucheta
-//---------------
-
-// technician-controller.getTechniciansAll and getTechniciansByAttribute
-app.get('/api/technician', (req, res) => {
-    for (const key in req.query) {
-        if (key ==='full_name') {
-            return res.json(technicians.filter(technician => technician.full_name === req.query[key]));
-        } else if (key ==='phone') {
-            return res.json(technicians.filter(technician => technician.phone === req.query[key]));
-        } else if (key === 'birthday') {
-            return res.json(technicians.filter(technician => technician.birthday === req.query[key]));
-        } else if (key === 'email') {
-            return res.json(technicians.filter(technician => technician.email === req.query[key]));
-        } else if (key === 'boilers') {
-            for (var i = 0 ; i < technicians.length; i++){
-                var found = technicians[i].boilers.some(boiler => boiler.$oid === req.query[key]);
-                if (found) {
-                    return res.json(technicians[i]);
-                }
-            }
-            if (!found) {
-                return res.json({msg: `Don't exist that boiler: ${req.query[key]}`})
-            }
-        } else if (key === 'types') {
-            for (var i = 0 ; i < technicians.length; i++){
-                var found = technicians[i].types.some(type => type.$oid === req.query[key]);
-                if (found) {
-                    return res.json(technicians[i]);
-                }
-            }
-            if (!found) {
-                return res.json({msg: `Don't exist that boiler: ${req.query[key]}`})
-            }
-        }
-        else if (key !== null) {
-           return res.json({msg:`Dont Exist that attribute: ${key}`});
-        }
-      }
-      res.json(technicians)
-});
-
-//technician-controller.getTechnicianById
-app.get('/api/technician/:id', (req, res) => {
-    const found = technicians.some(technician => technician._id.$oid === req.params.id);
-    if (found) {
-        res.json(technicians.filter(technician => technician._id.$oid === req.params.id));
-    } else {
-        res.json({msg:`No Technician with id: ${req.params.id}`});
-    }
-});
-
-//technician-controller.deleteTechnicianById
-app.delete('/api/technician/:id', (req, res) => {
-    const found = technicians.some(technician => technician._id.$oid === req.params.id);
-    if (found) {
-        res.json({msg:`Technician with id ${req.params.id} was deleted`, technicians: technicians.filter(technician => technician._id.$oid !== req.params.id)});
-    } else {
-        res.json({msg:`No Technician with id: ${req.params.id}`});
-    }
-});
-
 //------------------
 // Federico Morabito
 //------------------
